refactor: moving some build config to modules
