+++ app/build.gradle.kts
@@ -23,15 +23,7 @@ android {
         targetSdk = libs.versions.targetSdk.get().toInt()
         versionCode = libs.versions.version.code.get().toInt()
         versionName = libs.versions.version.name.get()
-
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-
-        // API keys and URLs
-        stringField(name = "API_URL", value = "https://api.themoviedb.org/3/")
-        stringField(name = "IMG_URL", value = "https://image.tmdb.org/t/p/w780")
-        stringField(name = "THUMBNAIL_BASE_URL", value = "https://img.youtube.com/vi")
-        stringField(name = "THUMBNAIL_QUALITY", value = "hqdefault.jpg")
-        // Build configurations
         buildConfigField(type = "boolean", name = "ADS_ARE_VISIBLE", value = "true")
         buildConfigField(type = "int", name = "PAGE_SIZE", value = "20")
     }
@@ -45,9 +37,7 @@ android {
             )
         }
     }
-    // Signing configurations
-    val activeSigning = System.getenv("OVER_ACTIVE_SIGNING") == "true"
-    if (activeSigning) {
+    if (isActiveSigning()) {
         // Signing configurations for different environments
         signingConfigs {
             create("prd") {
@@ -92,13 +82,13 @@ android {
             dimension = "version"
             applicationIdSuffix = ".homol"
             versionNameSuffix = "-hmg"
-            if (activeSigning) {
+            if (isActiveSigning()) {
                 signingConfig = signingConfigs.getByName(name = "hmg")
             }
         }
         create("prd") {
             setAppName(appName = "app_name_prd")
-            if (activeSigning) {
+            if (isActiveSigning()) {
                 signingConfig = signingConfigs.getByName(name = "prd")
             }
         }
@@ -113,9 +103,6 @@ android {
     buildFeatures {
         compose = true
         buildConfig = true
-    }
-    composeOptions {
-        kotlinCompilerExtensionVersion = "1.5.11"
     }
     hilt {
         enableAggregatingTask = true
@@ -173,7 +160,6 @@ dependencies {
 
     // Third-party libraries
     implementation(libs.timber)
-    // implementation(libs.retrofit)
     implementation(libs.converter.moshi)
     implementation(libs.logging.interceptor)
     implementation(libs.network.response.adapter)
@@ -186,7 +172,6 @@ dependencies {
     implementation(project(path = ":data"))
     implementation(project(path = ":domain"))
     implementation(project(path = ":presentation"))
-
 
     // Test dependencies
     testImplementation(libs.junit)
@@ -202,11 +187,12 @@ dependencies {
     debugImplementation(libs.androidx.ui.test.manifest)
 }
 
-// extensions
-fun VariantDimension.stringField(name: String, value: String?) {
-    buildConfigField(type = "String", name = name, value = "\"${value ?: ""}\"")
+private fun VariantDimension.stringField(name: String, value: String) {
+    buildConfigField(type = "String", name = name, value = "\"$value\"")
 }
 
-fun ApplicationProductFlavor.setAppName(appName: String) {
+private fun ApplicationProductFlavor.setAppName(appName: String) {
     resValue(type = "string", name = "app_name", value = "@string/$appName")
 }
+
+private fun isActiveSigning() = System.getenv("OVER_ACTIVE_SIGNING") == "true"
